Remove Kokkos_ThreadsExec_base.cpp file
